pyproject.toml
# go into a venv
# pip install build
# python -m build --sdist --outdir dist .

[project]
name = "rkwebutil"
description = "Rob Knop's Personal Web Utilities"
authors = [ { name="Rob Knop" } ]
maintainers = [ { name="Rob Knop", email="rknop@pobox.com" } ]
dynamic = [ "version" ]
requires-python = ">3.11"

classifiers = [
  # How mature is this project? Common values are
  #   3 - Alpha
  #   4 - Beta
  #   5 - Production/Stable
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]

dependencies = [
  "dateutils>=0.6.12,<0.7.0",
  "flask-session>=0.8.0,<0.9.0",
  "flask>=3.1,<4.0",
  "psycopg>=3.2.0,<4.0.0",
  "pycryptodome>=3.20,<4.0",
  "pytz>=2022.7.1",
  "pyyaml>=6.0.2,<7.0.0",
  "requests>=2.28,<3.0",
  "web.py==0.62"
]

license-files = [ "LICENSE" ]

[project.optional-dependencies]
test = [ "pytest",
         "remote-pdb"
       ]

[project.urls]
repository = "https://github.com/rknop/rkwebutil"


[build-system]
requires = [
  "setuptools>=62.1",
  "setuptools_scm[tomli]>6.2",
  "wheel"
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
namespaces = false
include = [ "rkwebutil" ]

[tool.setuptools.dynamic]
version = { attr = "rkwebutil.__version__" }


[tool.ruff]
line-length = 120
indent-width = 4
exclude = [
    ".git",
    ".github",
    ".ruff_cache",
    ".vscode",
    "__pypackages__",
    "docs",
]


[tool.ruff.lint]
preview = true
select = [ 'F', 'E101', 'E111', 'E112', 'E113', 'E115', 'E117',
           'E204', 'E223', 'E224', 'E242', 'E265', 'E273', 'E274', 'E275',
           'E301', 'E302', 'E305', 'E306', 'E401', 'E501', 'E502', 'E703',
           'E711', 'E713', 'E714', 'E72', 'E74',
           'W19', 'W29', 'W39', 'W605',
           'N804', 'N805', 'N807',
           'D206', 'D300', 'D301',
           'UP010', 'UP011', 'UP012', 'UP013', 'UP014', 'UP017', 'UP018', 'UP019',
           'UP02', 'UP030', 'UP031', 'UP033', 'UP034', 'UP035', 'UP036',
           'UP037', 'UP039', 'UP04',
           'PLE0100', 'PLE0101', 'PLE0116',
           'NPY',
           'RUF018',
           'E301', 'E302', 'E306', 'W505', 'D200', 'D212', 'RUF021' ]
